Python_Server/src/requirements.txt
pymysql==1.0.2
pycryptodome==3.12.0
django==4.2
django-cors-headers==3.10.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
requests==2.31.0
mysqlclient==2.2.7